+++ components/modify/EditFacilityDialog.tsx
@@ -13,7 +13,7 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import React, { useEffect, useState } from "react";
 import { toast } from "sonner";
-import { updateFacility } from "@/actions/facility";
+import { updateFacility } from "@/actions/admin/facility";
 
 interface EditFacilityDialogProps {
     isOpen: boolean;
